refactor: move `context.ts` file inside of `trpc` directory
